pyproject.toml
[project]
name = "pihole-audit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
